Implement V2 of Thingpedia Modules
In V2 of Thingpedia Modules, Channel classes are gone.
Instead, devices implement get_<query> and subscribe_<query>,
which either return a promise of a list or a stream with destroy()
method.

For compatibility with the existing Thingpedia, V1 is still
partially supported. In any case the major version of the Thingpedia
APIs has been bumped to indicate the break.