Switch from yarn from npm
Many years ago we moved from npm to yarn at a time when npm
was releasing version 5, which was very broken. Since then,
npm has gotten a lot more stable. Meanwhile, yarn has
accumulated unfixed bugs with git/github dependencies that
make it basically impossible to work with TypeScript modules,
because "prepare" scripts are not called correctly.

The tricky bit is that "npm" has no notion of "resolutions"
to force only a single version of a dependency. Instead,
we change the "thingtalk" dependency to be a peerDependency
here. This means that "thingpedia" will never have its own
copy of "thingtalk", and will always rely on somebody else
(like "genie-toolkit") to provide "thingtalk". And the world
will be happy, because there is only one "genie-toolkit"
and the almond-* do not depend directly on "thingtalk"