Migrate to ThingTalk 2.0
These are the small changes to adapt to the new version of
ThingTalk. The biggest change is that the Grammar module
is now called Syntax, and parseAndTypecheck does not exist.
The compat property "isMeta" is also gone, and the API of
Ast.Dataset changed a bit.

The rest is actual syntax changes in the test cases, and cosmetic
changes.