Make update logic simpler and more robust
Remove the on-disk cache for manifest files, and always query
Thingpedia. This is acceptable because querying Thingpedia is
fast, and it eliminates caching bugs.
(The on-disk cache of the actual device package is kept, and that
one is tied to the specified package_version)

Also, make it so that updateModule will keep using the old version
of a module if the new version has the same version number, so
we don't reload JS code unnecessarily.