Add interfaces to handle raw utterances in Thingpedia skills
For question answering/web search, we want to dispatch commands
to multiple backends and choose how to reply based on confidence.
To avoid putting third-party API code in the core Genie code,
define a set of Thingpedia interfaces that devices can implement.

For example, the Bing skill in Thingpedia will implement this
interface and ship every utterance to Bing (when enabled).