Add the ability to load secrets.json locally in development mode
When developing new skills that need API keys, those keys might
not be available in Thingpedia, until the skill is uploaded first.
This leads to very confusing errors for developers, and an ugly
workflow where one has to put the API key in the manifest, only
to remove them again when submitting for review.
With this, one sets the api_key=$? from the beginning, creates
a secrets.json in the same directory (encrypts with git-crypt),
and they're good to go.