Add a sqlite-based cache to the parameter provider
Parameter lists can be very big, and it's impractical to load them
in memory and scan them linerarly all the time. Instead, we can
use a small sqlite database to index the data and access it efficiently.

The database is updated automatically any time the TSV/JSON files
are updated.