Implement memory efficient sampling of parameters using sqlite indices
Instead of loading the entity parameter dataset in memory, we can
use sqlite queries to efficiently fetch specific rows. If we have
a lot of memory, the kernel will keep the database pages in memory
and it will be (almost) as fast, but we if we don't, we'll read them
from disk again instead of dying.